pyproject.toml
[project]
name = "dataagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=1.7.2",
    "kaggle>=2.0.0",
    "lxml-html-clean>=0.4.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "requests-html>=0.10.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
